Refactor RenderSystem into modular core, drawing, and utils components
